Running instrumented tests as part of daily pipeline (#215)
Adding a job to run instrumented tests as part of daily pipeline run.
We have good set of tests that run as part of instrumented tests, making
them part of daily pipeline run, so that we can pay close attention to
when something starts failing in these tests.

The new job will run instrumented tests for Common. Broker, Adal and
Msal
Here is a sample run:
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1108674&view=results

Also commenting out some of the oneauth related projects from
android-complete as we or oneauth team don't use them regularly as part
of android-complete.