Remove PROD brokerhost step, push RC Dist version to feed (#227)
This Pr fixes the Prod Brokerhost step that caused some issues last
release.

Instead of attempting to generate prod brokerhost in every run, we
instead push the RC Dist version to feed with the RC broker version used
as the feed version. This means we can simply pull it in the next run as
the PROD brokerhost. This way, we avoid the issue of using the incorrect
broker branch while building with PROD versions (this caused an issue in
the past in case changes were made in dev that didn't align with the
PROD versions).

Succesful release pipeline run

https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1126084&view=results