Update one auth integration branch (#232)
**What** : We are now triggering OneAuthTestApp pipeline from Android
daily build pipeline. We are also passing the latest common version to
this pipeline. But if the common version has any breaking changes, it
would break the oneAuth test app pipeline run. The fixes for these
breaks would be made in a special integration branch used in this PR.
And we would always use this integration branch to trigger the
oneAuthTestApp pipeline

Validation : Validated that we are able to pass the new common version
and build OneAuthTestApp successfully
- Android CI pipeline :
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1129979&view=results
- OneAuthTestApp pipeline :
https://office.visualstudio.com/OneAuth/_build/results?buildId=21212293&view=results