Add an extra UI Automation step to enable running LTW tests  (#250)
Run LTW UI Tests as part of daily pipeline. Also added a prefix to the
name of the test result to distinguish if the failing tests are related
to LTW or not.

MSAL_LTW_Branch and Broker_LTW_Branch are variables in the daily
pipeline that can be used to denote the LTW integration branches to
enable running LTW tests.