Fix step where oneauth test app is pulled from OneAuth pipeline (#284)
A PAT recently expired in one of the existing oneauth connections. The
connection was created by tanmay, but i don't have the permissions to
update it. I've created a new connection, in which i've granted admin
roles to multiple people in project and organization level, this makes
it easier to edit in the future.

New connection:
https://identitydivision.visualstudio.com/Engineering/_settings/adminservices?resourceId=cb1685ed-b09b-4a08-8ee0-44793edc38c3

Validation run:
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1225673&view=results