Update OneAuth integration branch to fix an issue with latest common (#289)
We recently made a change in common
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2146
that broke OneAuth pipeline. Matthew has made some changes to fix it by
updating their compileSDK version to 34 and a bunch of other changes in
his branch. This PR simply points to his branch instead of an old
integration branch which is not required anymore.