Skip LTW ui automation on scheduled days where it wouldn't pass (#303)
Adding this so we skip LTW ui automation on days where it would not pass
(not passing in broker selection and discovery flags) during scheduled
runs. this should reduce some cost over time.

**How:**
This is done by introducing a new output variable in the setup stage,
which will be set to false on Teusdays, Thursdays, and Saturday,
blocking LTW ui automation run on those days, when the necessary flags
are not being passed.

If it's a schedule run on monday, wednesday, or friday, or it's a manual
run, this new variable will not have any impact. This only affects
scheduled runs.