Switch LTW and AuthApp build steps to using Service Connection rather than PAT (#325)
Follow up to PR using WIF Service Connection for One Auth Test App step:
https://github.com/AzureAD/android-complete/pull/319

In this PR, we move away from using PATs for Authenticator and LTW build
stages to instead use the more secure WIF Service Connection.

Service Connection:
https://identitydivision.visualstudio.com/Engineering/_settings/adminservices?resourceId=e63aa87f-c686-42fa-abca-e4ed3289cab3

WIF Service Connection works by getting an access token from an
underlying Managed Identity (MI). The MI is given permissions to access
specific resources in other ADO organizations, which we use to run our
daily validation. For this to be possible, there are a set of
requirements:
1. The MI must be materialized in the external organizations with at
least "Basic" Access Level (This can be done by completing step 2 first,
and when the MI is used for something, it will automatically materialize
in the organization level. This may not be the case for every org
though)
2. The MI must be materialized in the partner project in the external
organization
3. The MI must have "View project-level information" permission granted
to it at the project level. This allows the Managed Identity to actually
see the resources it will be granted permissions to
4. Finally, the MI needs to have permissions granted to the resources in
that project. In our use case, that would be the build pipelines.

Here are some screenshots of what that looks like. The process for
getting these permissions granted is dependent on the organization, a
good place to start is the msado-dri@microsoft.com service alias.

Org-level materialization:

![image](https://github.com/AzureAD/android-complete/assets/88730756/7bbabe41-0ed3-4186-9f27-5e8c5ad0f2dc)

Project-level materialization:

![image](https://github.com/AzureAD/android-complete/assets/88730756/b22c8e05-1d76-4446-bf78-8ca1293c9199)

Resource permission granting:

![image](https://github.com/AzureAD/android-complete/assets/88730756/97b2ea31-31d9-4acc-b73c-a1779fef2675)

![image](https://github.com/AzureAD/android-complete/assets/88730756/4bb7de2a-99fe-432b-85ee-0004a3efff3b)

One limitation of WIF Service Connection, is it seems to not work for
accessing external artifact feeds. To solve this, i've created upstream
sources for our project's feed
(https://identitydivision.visualstudio.com/Engineering/_artifacts/feed/AndroidADAL)
that points to the feeds where LTW and AuthApp/CP artifacts are pushed
to. This removes the need for a service connection to access the
artifact feed, since our feed will have access to their artifacts, and
the pipeline can use build credentials to access them.

I'll need to do one more PR to add this solution for CP, as that is
waiting on security review.

Validation:

https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1316848&view=results

https://dev.azure.com/IdentityDivision/Engineering/_build/results?buildId=1316850&view=results