Move Cert Template in the daily pipeline (#338)
Not sure if this will work, as there's no way to test this considering
the nature of the bug.

Daily pipeline is not running instrumented tests as part of scheduled
runs or manual runs, but only when running on dev branch. It seems to be
related to CodeQL, which is automatically injected when running on dev,
but no other branch. So we won't know if this fixes the bug, until it's
merged.