Copilot instructions for github copilot, Fixes AB#3429068 (#381)
**Problem 1** : Today, github copilot is sometimes unable to understand
that the workspace contains sub repos and only picks the answer from one
of the modules and does not understand the whole picture.

**Solution** : I gave it an overview of all the repos, how data flows
between each module and also some overview of OneAuth library. Also gave
it an example question and how it should investigate such questions.

**Problem 2** : As of now, most of us write kusto queries manually and
do manual investigation when data from dashboards is not sufficient for
analysis.

**Solution** : Add kusto MCP server tool to github copilot in VSCode. I
have added instructions on what cluster, databse it should connect it,
what are our important tables, spans and what kind of investigations we
normally do and how it should proceed with those queries.

TODO : Write a TSG on how to add MCP server in VSCode. It is not
supported in Android Studio yet.


Fixes
[AB#3429068](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3429068)